requirements.txt
streamlit~=1.24.0
langchain~=0.0.142
openai
chromadb
tiktoken
streamlit-chat
python-decouple